feat: correct where functions live
